add practice in java

+++ Java/add1.java
@@ -1,10 +1,10 @@
 public class add1 {
-    public static void main (String[] args){
-        int a = 5;
+    public static void main(String[] args) {
+        byte a = 5;
         int b = 9;
         int c = 4;
         int d = a + b + c;
         System.out.println("Addition is: " + d);
     }
-    
+
 }
